게시판 - 로그인

+++ spring_6board/src/main/java/org/iclass/board/controller/CommunityController.java
@@ -10,7 +10,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.SessionAttribute;
-import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import lombok.AllArgsConstructor;
@@ -19,7 +18,6 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 @Slf4j
 @Controller
-@SessionAttributes(names = { "username" })
 public class CommunityController {
 	private CommunityService service;
 
@@ -61,7 +59,7 @@ public class CommunityController {
 	// 글 수정
 	@GetMapping("/community/modify")
 	public String modify(int idx,
-			@SessionAttribute(name = "username") String username,
+			@SessionAttribute(required = false) String username,
 			Model model) throws IllegalAccessException {
 		CommunityDTO dto = service.read(idx, false); // 글 수정은 조회수 카운트 안함
 		// 글번호 idx 의 작성자와 세션의 username 과 동일한지 비교
@@ -93,7 +91,12 @@ public class CommunityController {
 
 	// 글 삭제
 	@GetMapping("/community/remove")
-	public String remove(int idx, RedirectAttributes reAttr) {
+	public String remove(int idx,
+			@SessionAttribute(required = false) String username,
+			RedirectAttributes reAttr) throws IllegalAccessException {
+		if (!service.read(idx, false).getWriter().equals(username)) {
+			throw new IllegalAccessException("잘못된 접근입니다.");
+		}
 		service.remove(idx);
 		return "redirect:list";
 	}
